Add naive implementation

+++ kernels.cuh
@@ -1,3 +1,3 @@
 #pragma once
 
-extern "C" void twoSum(float* data, int* out, int target, int data_num);
+extern "C" void twoSum(int* data, int* out, int target, int data_num);
